gmail-client/pyproject.toml
[project]
name = "gmail-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "google-api-python-client>=2.188.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "mcp>=1.25.0",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
